# swagger.yaml
openapi: 3.1.0
info:
  title: BrReCoAPI
  version: '1.0'
  description: 'An ecommerce API. That allows users to register, login, create and save carts, view and add inventory, and place and track orders. The endpoints /auth/register- to create new user records, auth/login- for existing to log in, /user- to view or update user information, /carts- to create/modify user cart objects, /store- to view/add/edit product records, /orders- to place and view user order records'
  contact:
    name: J. Holt
    email: jordanh316@gmial.com
servers:
  - url: 'http://localhost:8081'
    description: Development/Local
paths:
...
  '/store/:productId/{productId}':
    get:
      summary: Get Product By ID
      tags:
        - GetProductById
      responses:
        '200':
          description: Product Record Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      operationId: 'get-store-:productId'
      description: Retrieves product record identified by productID parametet
    parameters:
      - schema:
          type: string
        name: productId
        in: path
        required: true
        description: Id of product record
    put:
      summary: Update A Product Record
      operationId: 'put-store-:productI'
      responses:
        '200':
          description: OK
      description: Find and update a user record by id
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - properties:
                    productName:
                      type: string
                - properties:
                    quantity:
                      type: integer
                - properties:
                    description:
                      type: string
                - properties:
                    price:
                      type: string
              type: object
        description: Body To Update A Product
      tags:
        - UpdateProductRecord
    delete:
      summary: Delete product record
      operationId: 'delete-store-:productId-productId'
      responses:
        '200':
          description: OK
      description: Find a product record by id and delete it
      tags:
        - DeleteProduct
...
components:
  schemas:
...
    Product:
      description: ''
      type: object
      x-examples:
        example-1:
          id: 708252b2-b107-4faa-84b0-f56ca2b39533
          vendorID: 29817d7a-abbf-4523-815e-bc6dd7344f8d
          productName: Bentley Super High Tech Gaming RollUp Phone
          description: The latest and fastest laptop on the market. Bentley Guarenteed
          price: '150000'
          quantity: '20'
          category: 'null'
          updatedAt: '2022-02-18T01:23:17.684Z'
          createdAt: '2022-02-18T01:23:17.684Z'
      properties:
        id:
          type: string
          minLength: 1
        vendorID:
          type: string
          minLength: 1
        productName:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        price:
          type: number
        quantity:
          type: integer
        category:
          type: array
          items:
            type: string
        updatedAt:
          type: string
          minLength: 1
        createdAt:
          type: string
          minLength: 1
      required:
        - id
        - vendorID
        - productName
        - description
        - price
        - quantity
        - updatedAt
        - createdAt
      examples:
        - id: string
          vendorID: string
          productName: string
          description: string
          price: 0
          quantity: 0
          category:
            - string
          updatedAt: string
          createdAt: string
...
  securitySchemes: {}
...